refactor(materializer): extract conversation-dispatcher (file #3)
Lift the cheapest-path dispatch (materializeCachedConversation) into
conversation-dispatcher.ts; the readmodel imports it. Materializer file is now
the public-API spine (readmodel + entry + scene assembly). Net 74/74, check 0.
