+++ packages/desktop/src/lib/acp/session-state/agent-panel-graph-materializer.ts
@@ -128,6 +128,7 @@ import {
 	materializeConversation,
 	materializeConversationWithOperationIndex,
 } from "./conversation-rebuild.js";
+import { materializeCachedConversation } from "./conversation-dispatcher.js";
 import {
 	interactionSceneEntryId,
 	materializeTranscriptEntry,
@@ -162,89 +163,6 @@ export function findLatestLiveAssistantEntry(
 	return null;
 }
 
-
-function materializeCachedConversation(
-	previous: CachedConversationState | null,
-	input: CachedConversationInput
-): CachedConversationState {
-	if (canReuseConversation(previous, input)) {
-		return previous;
-	}
-
-	if (canReuseConversationEntriesWithUpdatedActivity(previous, input)) {
-		const blockingInteractionRetarget = materializeBlockingInteractionRetargetConversation(
-			previous,
-			input
-		);
-		if (blockingInteractionRetarget !== null) {
-			return blockingInteractionRetarget;
-		}
-		const cachedConversation = previous as CachedConversationState;
-		return {
-			...cachedConversation,
-			activity: input.graph.activity,
-		};
-	}
-
-	const operationPatched = materializeOperationPatchedConversation(previous, input);
-	if (operationPatched !== null) {
-		return operationPatched;
-	}
-
-	const streamingStatePatched = materializeStreamingStatePatchedConversation(previous, input);
-	if (streamingStatePatched !== null) {
-		return streamingStatePatched;
-	}
-
-	const transcriptArrayPatched = materializeTranscriptArrayPatchedConversation(previous, input);
-	if (transcriptArrayPatched !== null) {
-		return transcriptArrayPatched;
-	}
-
-	const transcriptPatchedAndAppended = materializeTranscriptPatchedAndAppendedConversation(
-		previous,
-		input
-	);
-	if (transcriptPatchedAndAppended !== null) {
-		return transcriptPatchedAndAppended;
-	}
-
-	const transcriptPatched = materializeTranscriptPatchedConversation(previous, input);
-	if (transcriptPatched !== null) {
-		return transcriptPatched;
-	}
-
-	const transcriptTruncated = materializeTranscriptTruncatedConversation(previous, input);
-	if (transcriptTruncated !== null) {
-		return transcriptTruncated;
-	}
-
-	const transcriptAppended = materializeTranscriptAppendedConversation(previous, input);
-	if (transcriptAppended !== null) {
-		return transcriptAppended;
-	}
-
-	const interactionPatched = materializeInteractionPatchedConversation(previous, input);
-	if (interactionPatched !== null) {
-		return interactionPatched;
-	}
-
-	const operationIndex = buildOperationIndex(input.graph.operations);
-	const conversation = materializeConversationWithOperationIndex(input.graph, operationIndex);
-	return {
-		transcriptEntries: input.graph.transcriptSnapshot.entries,
-		operations: input.graph.operations,
-		operationIndex,
-		interactions: input.graph.interactions,
-		interactionById: buildInteractionIndex(input.graph.interactions),
-		turnState: input.graph.turnState,
-		activeStreamingTail: input.graph.activeStreamingTail,
-		activity: input.graph.activity,
-		transcriptEntryById: buildTranscriptEntryIndex(input.graph.transcriptSnapshot.entries),
-		conversation,
-		sceneEntryRowIndex: buildSceneEntryRowIndex(conversation.entries),
-	};
-}
 
 function materializeAgentPanelSceneFromConversation(
 	input: AgentPanelGraphMaterializerInput,
